+++ lockfile/lockfile-to-pnp/test/index.ts
@@ -1,3 +1,4 @@
+// cspell:ignore haspeer
 import path from 'path'
 import { lockfileToPackageRegistry } from '@pnpm/lockfile-to-pnp'