+++ cli/cli-utils/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @pnpm/cli-utils
 
+## 2.0.24
+
+### Patch Changes
+
+- Updated dependencies [ac5abd3ff]
+- Updated dependencies [b60bb6cbe]
+  - @pnpm/config@20.0.0
+  - @pnpm/default-reporter@12.4.3
+
 ## 2.0.23
 
 ### Patch Changes