+++ fetching/git-fetcher/CHANGELOG.md
@@ -1,5 +1,26 @@
 # @pnpm/git-fetcher
 
+## 12.0.7
+
+### Patch Changes
+
+- @pnpm/prepare-package@5.0.15
+
+## 12.0.6
+
+### Patch Changes
+
+- @pnpm/prepare-package@5.0.14
+- @pnpm/worker@0.3.5
+
+## 12.0.5
+
+### Patch Changes
+
+- Updated dependencies [08b65ff78]
+  - @pnpm/worker@0.3.4
+  - @pnpm/prepare-package@5.0.13
+
 ## 12.0.4
 
 ### Patch Changes