+++ packages/plugin-commands-init/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @pnpm/plugin-commands-init
 
+## 3.0.24
+
+### Patch Changes
+
+- Updated dependencies [ac5abd3ff]
+- Updated dependencies [b60bb6cbe]
+  - @pnpm/config@20.0.0
+  - @pnpm/cli-utils@2.0.24
+
 ## 3.0.23
 
 ### Patch Changes