+++ reviewing/plugin-commands-licenses/CHANGELOG.md
@@ -1,5 +1,26 @@
 # @pnpm/plugin-commands-licenses
 
+## 2.1.0
+
+### Minor Changes
+
+- fff7866f3: The `pnpm licenses list` command now accepts the `--filter` option to check the licenses of the dependencies of a subset of workspace projects [#5806](https://github.com/pnpm/pnpm/issues/5806).
+
+### Patch Changes
+
+- Updated dependencies [fff7866f3]
+  - @pnpm/license-scanner@2.1.0
+
+## 2.0.30
+
+### Patch Changes
+
+- Updated dependencies [ac5abd3ff]
+- Updated dependencies [b60bb6cbe]
+  - @pnpm/config@20.0.0
+  - @pnpm/license-scanner@2.0.22
+  - @pnpm/cli-utils@2.0.24
+
 ## 2.0.29
 
 ### Patch Changes